RPP: fix incorrect Interactions timeline
This is the same fix as crrev.com/c/6969008 but applied to the
UserInteractionsHandler, which does not use our generic event pairing
implementation (it should, but it works slightly differently with how we
construct the events, so I will tackle that in a separate CL).

Perfetto changed to reuse IDs for non-overlapping event pairs, and the
UserInteractionsHandler did not account for this. This CL ensures that
it does, meaning that we no longer pair events inaccurately in the
timeline.

Fixed: 446590230
Change-Id: I99571cbf131ab2b694eb0a7883f4f83a06fd9cb3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6972018
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>