Make min check in eventIsInBounds exclusive
I found a bug that resulted in `AICallTree.fromEvent` selecting a
different event for the "rootNode.event" than the input selected task.
Rarely, a super tiny event is associated with the call tree generated
from a task, rather than the expected long RunTask event. That made the
Performance AI agent wrongly use the short event when describing this
work.

It turns out that there was another task just before the one of interest
whose ts + dur exactly equalled the ts of the task the function cared
about. So it got included in the "overlapping events" array, and since
it occurs earlier than the actual long task, it was used as the root
node's event.

Now that we know a task can begin in the same microsecond as the
previous task ended, it seems correct to modify the min check in
eventIsInBounds from inclusive to exclusive.

Bug: none
Change-Id: Ib0119e132df8c6cff38f3444fe0c13c9d2df1c59
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6976655
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Paul Irish <paulirish@chromium.org>