[AiCompletion] Fix teaser showing up even after continued typing in sources
This was happening in this order:
* User types very fast "abc" and stops for 200ms
* Then, the debounced method is called and cleared
the timeout & triggered showing the teaser.
* Before the rendering timeout is passed, user continues
typing and types very fast, so the debounced method
is not called and the timeout is not cleared.

We need to clear the rendering timeout whenever the user
continues typing so that the rendering only happens if the
user waited in the same position.

Fixed: 447107739
Change-Id: Ifae955c6240df00a865a4444e8bb2abdb1a7fc09
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6979264
Reviewed-by: Samiya Caur <samiyac@chromium.org>
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Samiya Caur <samiyac@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>