Zoom into an event when its title in clicked
When using search, it is can be annoying to manually zoom into the event to see it in the flamechart. This adds a zoom when the event title is clicked in the summary pane.

Video: http://screencast/cast/NDc0NDE0NTM1MTQ3NTIwMHwyYTMzZmY0YS03Nw

Bug: None
Change-Id: I45f105fe1c6956c68c683ccb5085e9c0b10a409a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6938407
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>