Support legacy recipe versions
When updating recipes we must always make sure we do not impact CI and CQ on branches. Here we introduce a mechanism for allowing bots to run with older recipes when necessary.

When performing breaking changes on recipes we will update the legacy_recipe structure from definitions.star to allow branches to continue working with the old recipe.

Bug: 447563306
Change-Id: If27feb1cdbc6dc1f8dcb06960b109faeb1483bc3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6994300
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Liviu Rau <liviurau@chromium.org>