[GdpIntegration] Handle HTTP errors gracefully
Previously, if an HTTP error happened on `getProfile`
calls, we were assuming it is a user with no GDP
profile and was enabling the starter flows (e.g.
starter badge, and the "Sign up" button in Settings).

However, considering that an overloaded API can
provide 503 responses much more frequently, we need
to handle those failures gracefully and disable
the starter flows.

There are 2 types of error handling here:
1. Internal to `GdpClient`: `makeHttpRequest`
throws a `GdpError` for non-200 responses and
internally, calls in the `makeHttpRequest` handled
in try-catch blocks.
2. External to `GdpClient`: For the external methods
(like `getProfile` or `createAward`), if the request
succeeds we return the response otherwise we return
`null` to the caller.

Fixed: 448012975
Change-Id: I910542a4edb13999b82e0c91d7c7aef26456adb8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6992374
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>