Tooltip: ensure "ESC" always closes the tooltip.
An a11y review of the "Export trace" dialog found that ESC does not
close the tooltip, which it should do. We had the logic for doing this
already in the code, we just need to activate the ESC logic regardless
of what mode (use click, use focus or use hotkey) the tooltip is in.

I also updated the Export trace dialog to listen to ESC and close.

R=alinavarkki@chromium.org

Bug: 447629421
Change-Id: I439ef73805ab41693f2102ec6dcbd35d77abda72
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6999769
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>