Select the primary page target on added in the OutermostTargetSelector.
The target page dropdown menu now automatically selects the added page
if the page is the primary page.

This fixes the problem that the embedder-triggerd prerendering page can
be chosen when the current page is reloaded as there is a timing where
the prerendered page can be only the existing page.

Change-Id: If2f25fa674a1e7b61e9b78cbef00cb7f62921138
Bug: 443060872
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6997978
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>