[AI Assistance] Introduce dedicated function for styling
New function called `getStyles` returns computed and authored styles at
the same time that makes it easy for the LLM to get relevant style
information for any specific style.

The CL changes the element descriptions to include UIDs so that the LLM
can pass them back to `getStyles`. This is aligned with the approach we
tested in the MCP server.

Bug: 435123594
Change-Id: Ia7f8e7e1e7bfd586394221fd3d5b56521db2aa1e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6930093
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>