RPP: add number formatting to Counters Graph
We were formatting numbers with a space; but that is odd and not common
for many locations where DevTools is used. This CL changes the formatter
to an Intl.NumberFormat so we pick an accurate format based on the
user's locale.

Screenshot for Googlers: screen/AraQxeaeEUfYHkL

Bug: 448204885
Change-Id: I67cd27941be85f2f162eefaae41c209a82fce869
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7002540
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>