Remove flaky test due to little value it has :-)
The test caused tree closure recently due to it timeouting.
Locally it is several times slower than other tests in the suite due to
amount of actions it takes.
On the other hand, it doesn't test behavior of the sources panel,
but only click helper which should be tested separately.

Bug: 448586593
Change-Id: I145aef0bb96cac336c22bd6f21d20a767ee238ac
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7002659
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Piotr Paulski <piotrpaulski@chromium.org>