PerfAI: pick no_navigation Insight set if it's the only choice
I noticed this bug when wondering why the AI had no idea about the
interaction I was asking it to debug. The code to select the first
insight assumed we'd find one with a navigation, but often for
interactions that is not the case. This CL updates the logic to:

1. If there is only 1 InsightSet, use that.
2. If there are >1, prefer the first with a navigation attached, but
   otherwise fallback to the first set otherwise.

This fixes the behaviour where for no-navigation traces the AI has no
idea about any insights.

(Long term we should still do the proper fixing where we tell the AI
about all of them).

R=cjamcl@chromium.org

Fixed: 449113697
Change-Id: I806ca7fa7770845c08b0d29afa3b8f484cb2793a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7003546
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>