Fix console-context-selector_test's flakiness
The test was failing around 80% of the time for me (locally) waiting for
the menu that wasn't there.
The menu was disappearing due to ConsolePanel still updating its state
in reaction to changes in SourcesPanel.
To fix the issue, I reimplemented the test to click the menu again if
it does not open as expected.

Fixed: 430774159
Change-Id: If7315b1d263f98f915bf31c7f8cc4704e3f6e2aa
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7008372
Commit-Queue: Piotr Paulski <piotrpaulski@chromium.org>
Auto-Submit: Piotr Paulski <piotrpaulski@chromium.org>
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>