Add ARIA label and placeholder to quota override editor in StorageView
Before: Narrator wasn't announcing the checked state of the 'simulate
custom storage quota' checkbox. This was because the quotaOverrideEditor
would steal focus right away after the checkbox is toggled.

After: We avoid setting the focus, and include placeholder and
aria-label for the quotaOverrideEditor input so the screen reader still
announces the input label when the user navigates to it.

Bug: 448680444
Change-Id: Ie2bd69439b96c7935879af4add46edfbbcb71d34
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7003739
Commit-Queue: Sylvester Elorm Coch <elormcoch@microsoft.com>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>