[console] Fix holding key down changing history entry
Currently if a user has big snippet and tried to go to the top or bottom the history is update with the respective one.

A better UX is that we stop if they were holding the button down.
If we are navigating the history this work as before.

Fixed: 449722141
Change-Id: Icd3ffbe39e30f86b924a825494947cb9b5c4e900
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7006352
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Finnur Thorarinsson <finnur@chromium.org>