Fix for JSONView not updating when selecting another report
After a previous CL (https://crrev.com/c/6859315), the content of the
JSONView did not update when `input.focusedReport` changes. To fix
this, the widget needs to have a setter method for updating the JSON.

This CL creates a child class of `SearchableView` which has such a
setter.

Bug: 407750294
Change-Id: I6bc59379d883bbf2a4e6bc87c0a2b91845ce4e5d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7000090
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>