Improve console-context-selector_test by merging two helpers into one
Modify `clickMenuAndWaitForItemWithText` to be able to search for
multiple menu items, thus removing the need for separate helper
that only looks for menu items within open menu.
`waitForMenuItemWithText` is inferior and caused flakiness in the past
due to its inability to react to menu being closed due to background
processing in the panel.

Bug: 430774159
Change-Id: I12de704170d916f4ef094882de1a638e0c792fc1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7013470
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Piotr Paulski <piotrpaulski@chromium.org>