Use tree scopes to resolve the context step when checking property overrides
In UA shadow dom, standards-track pseudo-elements (e.g.,
::details-content) can override style attribute style. These rules win
in the context step of the cascade since both styles are considered to
be from the author origin. To resolve this, we can look up the
originating tree scopes when checking overrides and apply the correct
one depending on which one is from a higher context.

This does not yet fix crbug.com/429993394, since the :host rule is
attributed to the shadow's tree scope.

Fixed: 40192802
Change-Id: I3880ccbcdffca0197a0f5c5ee828ceba48a2bc19
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7018472
Reviewed-by: Eric Leese <leese@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>