[sdk] Wire up scope info from AST into the source map
This CL wires up the fallback AST-based scope information into the
source map.

It is slightly awkward: SourceMap and SourceMapManager were on
a generic "client" which in reality is either a SDK.Script or
an SDK.CSSStyleHeader. To be able to pass the SDK.Script to
the SourceMap, we allow creators of the SourceMapManager to
pass a SourceMap factory function. This way, the DebuggerModel
can pass the SDK.Script.

Drive-by: Hide new scope view also behind experiment, since we
now have SourceMapScopesInfo objects without proposal scopes.

R=pfaffe@chromium.org

Bug: 449575926
Change-Id: I26ba583e36102af293dd487eeb759766e456fce5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7021137
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>