Fix UI eng violations in TimelinePaintImageView
There is also TimelinePaintProfilerView in this file, which is a SplitWidget that might be harder to refactor since SplitWidget's core functionality is imperative: the primary job of a SplitWidget is to manage the size and position of its child elements. I will look it in another cl.

Bug: 407751342
Change-Id: Icf08145cdcb30e4ed9499fc7581435b3ddea6596
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6939304
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>