[RPP] Force tab targetType in trace_app
This is required for some features to work, such as linkifying the
console source location in the bottom-up / call-tree tables. The short
summary is that `primaryPageTarget()` does not return the correct target
for rehydrated connections unless the first / root target is of type
tab, which breaks linkifying.

Bug: 427455148
Change-Id: I62e656d9f92892d8b051e0c7b98c663b432499d1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7037208
Reviewed-by: Paul Irish <paulirish@chromium.org>
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>