[scopes] Introduce 'translateCallSite' on SourceMapScopesInfo
This method is a combination of "findInlinedFunctions" and
"expandCallFrame". These were split because the initial prototype
used the same mechanism as DWARF: Expand inlined frames first by
creating virtual frames. Then, during live location updates, figure
out the actual call-site for each of the "virtual" frames.

With the new StackTrace abstraction this is no longer required and
we can translate a non-outlined frame in a single go.

This is exactly what "translateCallSite" does, it takes a single
generated position and returns one or more translated frames.

The tests are adapted from "expandCallFrame".

R=pfaffe@chromium.org

Bug: 433162438
Change-Id: I697c48c7d1e8b9f496cb033a972506475ffbfed7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7053162
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>