[Console Insights Teasers] Minimize Tooltip repositioning
Ensure the height of the tooltip (mostly) stays unchanged in the
transition from loading state to answer state, because height changes
can cause the tooltip to be repositioned (e.g. from being on the
bottom of the anchor to being on top of the anchor).

- Render the footer unconditionally
- Clamp the heading to 1 line
- Clamp the response text to 4 lines
- Adjust the height of the placeholder SVG

Bug: 443618746
Change-Id: I00b5d18f7c6710d1bbf0afce92121f22a8c413f2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7069572
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>