[RPP] Calculate id only once per event in ungroupedTopNodes
This is a minor improvement to generating an id twice for each event (on
start and end). Instead, use a stack to calculate it just once.

I explored more serious improvements here related to avoiding the second
Map for totalTimeById – profiling showed that a significant amount of
time was spent in Map. However, it proved too difficult to figure out in
the time I set aside for this.

I did discover a potential logic error related to marking when an event
has children. I added a comment and a test about it, but I'm not sure
the impact is meaningful so I'll leave it there.

Bug: 444483828
Change-Id: Ibfa584d55306177acff81e18de00a5411a7c2ae5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6940959
Reviewed-by: Paul Irish <paulirish@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>