[Lighthouse] Tweak auto-focusing after generating report
In 2019, crrev.com/c/1974456 added a `focus()` call to this button in
order to fix a focus issue in DevTools after Lighthouse runs. The issue
was that the active element after a report is generated is otherwise the
first element in DevTools' main toolbar, rather than something in the
Lighthouse panel.

Instead of focusing the "new report" button (which is visible), set the
focus on the Lighthouse URL element (which has no visible focus styles).

Fixed: 445660552
Change-Id: Ia9481f0a8e606eec91a6205a7e87c77705179c79
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7067943
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>