Improve screen reader announcement when recording is started.
Issue: Screen reader is announcing unnecessary extra information when
user select the Start Recoding Events button in Background Fetch

This is because after the record button is clicked the focus is lost on
the emptyWidget which contains information about the recording state. We
should announce this information and keep the focus within the widget.

Bug: 454082774
Change-Id: Ieae12e925bc3cc4cecdf3d84a6314bd8da3d1203
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7069765
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Sylvester Elorm Coch <elormcoch@microsoft.com>
Reviewed-by: Mathias Bynens <mathias@chromium.org>