[protocol monitor] Manually map session IDs to targets
We plan to refactor the protocol client and add a layer that deals
with CDP messages only without handling sessions and targets. As such,
the functions sniffing on the raw CDP traffic, won't have the target
available in the future.

This CL changes the protocol monitor to manually maintain a map
between sessionID -> target so we can safely remove the 'target'
parameter from 'test.onMessageSent' and 'test.onMessageReceived'.

R=kimanh@chromium.org

Bug: 453469270
Change-Id: Idb71f713ff49526fc0d0af040304e9eb2aba79b4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7069516
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>