[Lighthouse] Prevent bad state when canceling very quickly
The main issue is that callers of startLighthouse/cancel did not and
cannot await their completion (they are button callbacks), so multiple
cancels or starts can be in-flight at the same time. To fix, I added a
simple "lastAction" promise on the controller to force the completion of
the last action before doing the next one.

I also ran this code through Gemini a bit. While it didn't help me find
the fix, it did restructure the code a bit better, so I kept the
changes. In particular, we now have a `LighthouseRun` class, whose
instances follow a run's lifecycle. We keep the LighthouseController
singleton.

Fixed: 40728672
Change-Id: I814a8809687db6ffa20c1804d2c5fb739af93ec3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7075872
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Auto-Submit: Connor Clark <cjamcl@chromium.org>