RPP: fix font family in screenshots
A fix we applied to the TimelineFlameChartView a long long time ago in
the age of interaction tests to pick the right font was not correctly
adapting to the Roboto font that gets set in test_setup.ts.

Curiously this did not replicate on my Cloudtop, but did reliably on the
bots, which caused these tests to be skipped.

I think in this new world we can greatly simplify here and just always
use the platform font, as we only run screenshots on Linux.

If this works and these tests are stable now, I will follow up with a
CL to remove `perf_ui/Font.ts` as we don't need that module now we just
use the host font.

Bug: 453626439, 453630637, 453629222
Change-Id: I6006cee53142e5bb8e0c9ae091490967c90dd360
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7079233
Reviewed-by: Peter Müller <petermueller@chromium.org>
Reviewed-by: Piotr Paulski <piotrpaulski@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>