Improve screen reader announcement of bfcache test result
Issue:
After the test bfcache button is invoked, the re-renders and so focus is lost from the button to the devtools window resulting in incorrect and missing screenreader announcements.

Fix:
After rendering the results, we announce the tests were completed and keep focus on the results section view where the announcement happens.

Bug: 455905510
Change-Id: I15674c8291a703317dccad7dd42d57d23e231a62
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7090941
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Commit-Queue: Sylvester Elorm Coch <elormcoch@microsoft.com>