[protocol] Promisify InspectorBackend.sendMessage
This CL changes 'sendMessage' to return a Promise instead of a
callback. The AgentPrototype one layer higher already does the same
so we just shift the whole promisifying down one layer.

This will make it easier to change over to a CDPConnection impl.

Drive-by: Inline the two "dispatch*Error" static methods.

The tests are rather sensitive to timing and the "mocked
by default" resource tree is also causing problems in some
cases. The CL puts some bandaids on these so we can land the
changes in `InspectorBackend.ts`.

R=alexrudenko@chromium.org

Bug: 453469270
Change-Id: Icb61cc5defbb0ae5193469c5a3ef0a63aa903a06
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7090211
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>