[docs] Build with typescript checks on Github
In order to make sure that the dependencies of our docs are
existent, we need to build with typescript checks. If a dependency
on the component is missing, typescript checks will throw an error
during build time when building the scripts/component_docs target.

Drive-by: Change to order of documentation navigation sidebar
Bug: 414332459
Change-Id: Ife52bcd35501af930e7685c4765bb9c9cf80fc74
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7100240
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>