[docs] Replace the container when reloading a new component doc
In our documentation we have a mixture of lit templates and the legacy
imperative way of creating the components. If we only reset the
`innerHTML` of the container, Lit isn't aware of the DOM changes, and
components may not be rerendered if navigated to again.

Instead, let's start off with a fresh container each time we navigate to
a new doc. This also minimizes the side effects on the container across
docs.

Bypass-Check-License: This doesn't change any licenses (but a parent does)
Bug: 414332459
Change-Id: Iadf924900f05c90f62d9b13ed3ed998a2161aa27
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7100478
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>