Mostly migrate RequestTimingView.ts
The widget uses an ObjectUI tree which is blocking full migration.

In addition to migrating, this CL adds highlighting and an icon for
throttled requests. This is added to the "Waiting" timing, despite
throttling being able to affect download throughput. The way throttling
is implemented, delays are only factored into the "Waiting" timing,
which is clearly visible so it makes no sense to put the highlighting
somewhere else.

Fixed: 444376779
Bug: 407751338
Change-Id: I5ef142d1a95a02e96706f553ca820df403a24b4c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7079297
Reviewed-by: Eric Leese <leese@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Philip Pfaffe <pfaffe@chromium.org>