[recorder] Introduce new puppeteer.ConnectionAdapter
This CL introduces an alternative to "parallel connection".

The key ingredients for this to work is the new DevTools
"CDPConnection". It has a promisified 'send' and arbitrary
listeners can listen to CDP events.

To proxy puppeteer traffic over the new CDPConnection, we
wrap the CDPConnection in a "PuppeteerConnectionAdapter"
which makes it look like a "puppeteer.Connection".

The only responsibility of the adapter is to filter out
CDPEvents based on the session ID: The adapter sees ALL events,
but it should only forward those for sessions that puppeteer
has actual sessions. The rest belongs to DevTools or Lighthouse.

R=alexrudenko@chromium.org

Bug: 453469270
Change-Id: I2924f5299bf5beb19ffceee33ff7c7b6b92e8451
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7105521
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>