[lighthouse] Use CDPConnection instead of parallel connection
Both on the frontend-side and the worker-side, this CL updates
Lighthouse to use the new CDPConnection.

On the frontend-side, we create a new CDP session and then use
the standard CDPConnection to send/receive messages, which
are forwarded nearly 1:1 to the worker.

On the worker-side, we create a custom transport with
a CDPConnection, and then wrap it with the new puppeteer
adapter.

Bug: 453469270
Change-Id: I737816f4ce657210079413ba888e9c26897dac82
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7106261
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>