TraceEngine: update render_blocking with preload status change
This CL fixes a bug where a blocking resource (e.g. stylesheet) is not
marked as blocking in RPP because it is first preloaded.

This was because we did not pay attention to the
PreloadRenderBlockingStatusChange event which exists for this exact
reason. With this CL when we construct the final synthetic network
request for a resource, we will use the render blocking status from the
last PreloadRenderBlockingStatusChange event (if we have any), before
falling back to the render blocking status on the SendRequest event.

R=cjamcl@chromium.org

Fixed: 457323832
Change-Id: I996a3372a4ec8f9cbf8d1c387b2ec44a80aa9c14
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7112841
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>