[node] Change node_app regression test
Once we wrap the transport in a CDPConnection, we won't have access
to the underlying transport.

To still verify that we send Node.js CDP traffic via NodeConnection,
we instead check that the parent's targetAgent got invoked.

We don't check the full CDP traffic, just that a Debugger.enable was
in there.

Bug: 453469270
Change-Id: I3338d2e4ef4286ddb32cbf5850098b7018fee0ea
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7123818
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>