[foundation] Add basic Universe, only instantiating settings for now
This CL adds the basic scaffolding for what will become bootstrapping
for DevTools foundation.

For now, we instantiate the global Settings instance as this will be
required for nearly everything else.

The CreationOptions are also not final. For now, they are used to
pass in anything "host platform" dependent, but we'll move these
behind better types and interfaces as the migration goes on. For now,
they are a grab bag of host-dependent dependencies required to
instantiate foundation classes.

The CL also adds a unit test running both in Node and the browser
to make sure we can instantiate a Universe in both.

Moving forward, we'll add a DevToolsContext so we can scope these
globals. We'll also have a single global DevToolsContext for
migration purposes.

Bug: 458180550
Change-Id: If1f1b172517ef4e8e3992764cc2f4cf811e24a14
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7124027
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>