[sdk] Don't log an error when a stream fails to close
Failing to close a stream is not recoverable: Either the stream closes
successfully anyway or it is in a broken state, which is equivalent
to being closed.

If the frontend needs to do handle this case, we should throw an error
and handle it in the caller, similar to other IOModel methods.

It seems the e2e test got unlucky enough that it consistently runs
into this on bots.

R=alexrudenko@chromium.org

Fixed: 456094035
Change-Id: Id11d03e1c60dcb39177ce0beeccf30d7733e2127
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7137439
Auto-Submit: Simon Zünd <szuend@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>