[common] Pass SettingRegistrations to Settings constructor directly
This is a purely mechanical change in that we lift the call of
'getRegisteredSettings' to the place where we create 'Settings'
instances.

As a nice bonus, calling 'settings.getRegisteredSettings()' will now
always return the set of registrations used to initialize that
particular instance, rather then the current global state.

R=pfaffe@chromium.org

Bug: 458180550,456121853,456126110
Change-Id: Ifb7f8f1357b4073b4521afed5b53f78523f1a69b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7137440
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>