[root] Introduce DevToolsContext and scope TargetManager
This CL adds a new class called `DevToolsContext` to root. It's
intended to hold instances that are currently global singletons so we
can have one per Universe.

For now, we have a single global instance and re-direct existing
`instance` accessors to it.

To show how this would look like, this CL changes the `TargetManager`
over.

Unfortunately we also need to add `has` and `delete` accessors for now
to keep existing tests working. So we basically store the instance
in a global `DevToolsContext` rather then a ES module variable, but
the rest is the same (for now).

R=alexrudenko@chromium.org, pfaffe@chromium.org

Bug: 458180550
Change-Id: I2093136ab028328fd692c5b9d23b38f978ffb4db
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7141478
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>