[docs] Migrate theme support docs to static documentation page
In addition to migrating, we remove the 'Dark Theme' column (which
hasn't been working as part of the component docs anymore either), as
the dark theme is nowadays dependent on the css styles of the root
class.

Also, remove the elevation part and focus on colors for now. In the
future we can think of documenting all --sys-(non-colors).

Bypass-Check-License: Only moving docs
Bug: 414332459
Change-Id: I2c53eea611fc49f1db80e8efb4b9bac4ad7f0ffc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7146266
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>