[testing] Add alternative to 'describeWithMockConnection'
This CL introduces a new MockCDPConnection class. It is intended to
work similar to the existing MockConnection.ts, but rather then mocking
on the 'ConnectionTransport' level globally, we mock on the
connection level on a per-target basis.

This means each MockCDPConnection only affects those targets
where it is passed to `createTarget`. Note that child targets
are the exception: They inherit the connection from the parent, but
the helper will throw early to warn the user in this case.

For easy migration, we automatically pass a MockCDPConnection with
0 hooks when creating targets outside of 'describeWithMockConnection'.

Another thing to note is that the `MockCDPConnection` errors on
every CDP method not-explicitly stubbed. This will require setting up
a set of "default empty handlers" as many auto start models trigger
CDP calls (e.g. `DOM.enable`).

R=alexrudenko@chromium.org, dsv@chromium.org

Bug: 451502260
Change-Id: I8d05496098a5169af0972dd9d6aabd52bd6df8bc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7146861
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>