[bindings] Set CSS- and DebuggerWorkspaceBinding on ResourceMapping
CSSWorkspaceBinding and DebuggerWorkspaceBinding form dependency
cycles with ResourceMapping. Since a ResourceMapping is already
passed via constructor to the *WorkspaceBindings, so we can't do
the reverse for the ResourceMapping constructor.

To get rid of global `instance()` accessors in ResourceMapping, we
add a pair of accessors that the CSS- and DebuggerWorkspaceBinding
use to install themselves on the ResourceMapping.

A follow-up CL will pass the ResourceMapping down into the actual
Bindings so we can replace `instance()` calls.

R=nvitkov@chromium.org, pfaffe@chromium.org

Bug: 458180550
Change-Id: I57e6cec91445a3472f54449c28e94b3fd180cb52
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7146228
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>