Add CrUX Vis link to live metrics view (perf panel landing page).
This adds a link to the CrUX Vis viewer
(https://cruxvis.withgoogle.com/) for field data to the live metrics
view.

To be consistent with PageSpeed Insights, the link is placed in
parenthesis after the collection period. E.g.,

Collection period: Sep 15, 2025 - Oct 12, 2025 (View more history)

Note that when the collection period begin / end aren't rendered, then
this "(View more history)" link also won't be rendered. Also, the actual
link will use the same "scope" that the user has made in the panel, that
is url-level vs. origin level data as well as device type.

To get this done I added a normalizedUrl field to the CrUXManager; this
makes it so that we use the same URL for CrUX vis that we're using to
make CrUX API requests.

Bug: 405467537
Change-Id: I63271c6eab7ba30c69b948a07954b09a6e807a8c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/7040531
Commit-Queue: Johannes Henkel <johannes@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Auto-Submit: Johannes Henkel <johannes@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>